Abstract trie node functionality on its own struct *TrieNode
